+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ### Unreleased
+* ++ Admin client. Added '.call_add()'.
 * ++ Admin client. Exposed 'url_listing' and 'url_add' attributes.
 
 ### v1.6.0 [2025-09-28]